testing/web-platform/tests/orientation-event/motion/window-property.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
'use strict';

function hasDeviceMotionEventProperty()
{
  for (let property in window) {
    if (property == "DeviceMotionEvent")
      return true;
  }
  return false;
}

function hasOnDeviceMotionProperty()
{
  for (let property in window) {
    if (property == "ondevicemotion")
      return true;
  }
  return false;
}

test(test => {
  assert_equals(typeof window.DeviceMotionEvent, 'function');
  assert_false(hasDeviceMotionEventProperty());
  assert_true('DeviceMotionEvent' in window);
  assert_true(window.hasOwnProperty('DeviceMotionEvent'));

  assert_equals(typeof window.ondevicemotion, 'object');
  assert_true(hasOnDeviceMotionProperty());
  assert_true('ondevicemotion' in window);
  assert_true(window.hasOwnProperty('ondevicemotion'));
}, 'Tests that the window.DeviceMotionEvent and window.ondevicemotion properties are present.');
</script>
